refactor: add translator module
