+++ src-tauri/src/services/mod.rs
@@ -1,4 +1,3 @@
 pub mod downloader;
 pub(crate) mod translator;
-mod processor;
 pub(crate) mod sniffer;
