+++ src-tauri/src/services/translator/processor.rs
